backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.2.0
mediapipe==0.10.14
tensorflow==2.13.0
keras==2.13.1
python-dotenv==1.0.0
